@@ -5,17 +5,17 @@
   "main": "index.js",
   "scripts": {
     "pretest": "npm run lint",
-    "test": "nyc mocha",
+    "test": "nyc mocha --require @babel/register",
     "test:debug": "mocha --inspect --debug-brk",
     "codecov": "cat ./coverage/lcov.info | codecov",
     "lint": "xo index.js test/test.spec.js",
-    "start": "mocha --watch",
+    "start": "mocha --require @babel/register --watch",
     "preversion": "npm test",
     "postversion": "npm publish"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/mattlewis92/htmlhint-loader.git"
+    "url": "git+https://github.com/thedaviddias/htmlhint-loader.git"
   },
   "keywords": [
     "webpack",
@@ -25,33 +25,37 @@
     "lint",
     "html"
   ],
-  "author": "Matt Lewis",
+  "author": "David Dias, Matt Lewis",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/mattlewis92/htmlhint-loader/issues"
+    "url": "https://github.com/thedaviddias/htmlhint-loader/issues"
   },
-  "homepage": "https://github.com/mattlewis92/htmlhint-loader#readme",
+  "homepage": "https://github.com/thedaviddias/htmlhint-loader#readme",
   "dependencies": {
-    "chalk": "^2.3.2",
-    "glob": "^7.1.2",
-    "htmlhint": "^0.9.7",
+    "chalk": "^2.4.1",
+    "glob": "^7.1.3",
+    "htmlhint": "^0.10.1",
     "loader-utils": "^1.0.2",
     "strip-bom": "^3.0.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0-rc.3",
+    "@babel/preset-env": "^7.0.0-rc.3",
+    "@babel/register": "^7.0.0",
     "chai": "^4.0.1",
     "codecov-lite": "^0.1.3",
-    "mocha": "^5.0.2",
-    "nyc": "^11.0.1",
+    "coffeescript": "^2.3.1",
+    "mocha": "^5.2.0",
+    "nyc": "^13.0.1",
     "raw-loader": "^0.5.1",
-    "sinon": "^4.4.2",
-    "sinon-chai": "^3.0.0",
+    "sinon": "^6.1.5",
+    "sinon-chai": "^3.2.0",
     "strip-ansi": "^4.0.0",
-    "webpack": "^4.1.0",
-    "xo": "^0.20.3"
+    "webpack": "^4.17.2",
+    "xo": "^0.23.0"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=6.0.0"
   },
   "files": [
     "index.js"
